feat(snap): config hook support for add-secretstore-tokens
This commit adds configure hook support for the security-
file-token-provider (launched by secretstore-setup) env var
ADD_SECRETSTORE_TOKENS which can be used to cause additional
vault tokens to be generated for application or device service
snaps. This setting is meant to be used in conjunction with
the edgex-secretstore-token content interface.

Signed-off-by: Tony Espy <espy@canonical.com>