fix(snap): disable Vault TLS in post-refresh hook
EdgeX assumes Vault doesn't use TLS. It is disabled in a Vault config file
installed in the install hook. When refreshing from an earlier version
of EdgeX, we may have an invalid config file with TLS enabled.

This update therefore uses the post-refresh hook to update the
Vault configuration file.

Signed-off-by: Siggi Skulason <siggi.skulason@canonical.com>